test(lotto): ValidationError 테스트 파일 위치 수정 및 개별화
